progress stopped

+++ Gordon360/Controllers/MembershipsController.cs
@@ -35,26 +35,27 @@ namespace Gordon360.Controllers
         [StateYourBusiness(operation = Operation.READ_PARTIAL, resource = Resource.MEMBERSHIP)]
         public ActionResult<IEnumerable<MembershipView>> GetMemberships(bool? myProf, string? involvementCode = null, string? username = null, string? sessionCode = null, [FromQuery] List<string>? participationTypes = null)
         {
+            var authenticatedUserUsername = AuthUtils.GetUsername(User);
+            var viewerGroups = AuthUtils.GetGroups(User);
+            bool mp = myProf ?? false;
+
             var memberships = _membershipService.GetMemberships(
                 activityCode: involvementCode,
                 username: username,
                 sessionCode: sessionCode,
                 participationTypes: participationTypes);
 
-            if (username is not null)
+            // SiteAdmin and Police can see all of anyone's memberships. If looking at your public profile should not show up private involvements.
+            if ((viewerGroups.Contains(AuthGroup.SiteAdmin)
+                || viewerGroups.Contains(AuthGroup.Police)
+                ) )
             {
-                var authenticatedUserUsername = AuthUtils.GetUsername(User);
-                var viewerGroups = AuthUtils.GetGroups(User);
-                bool mp = myProf ?? false;
-
-                // User can see all their own memberships. SiteAdmin and Police can see all of anyone's memberships
-                if (!(username == authenticatedUserUsername
-                    || viewerGroups.Contains(AuthGroup.SiteAdmin)
-                    || viewerGroups.Contains(AuthGroup.Police)
-                    ) || !mp)
-                {
-                    memberships = _membershipService.RemovePrivateMemberships(memberships, authenticatedUserUsername);
-                }
+                memberships = _membershipService.RemovePrivateMemberships(memberships, authenticatedUserUsername);
+                return Ok(memberships);
+            }
+            // User can see all their own memberships.
+            if ((username is not null) && !(username == authenticatedUserUsername)){
+                memberships = _membershipService.RemovePrivateMemberships(memberships, authenticatedUserUsername);
             }
 
             return Ok(memberships);
